+++ lib/postgres.dart
@@ -21,7 +21,7 @@ export 'src/types/geo_types.dart';
 export 'src/types/range_types.dart';
 export 'src/types/text_search.dart'
     show TsVector, TsWord, TsWordPos, TsWeight, TsQuery;
-export 'src/types/type_registry.dart' show TypeRegistry;
+export 'src/types/type_registry.dart' show SqlNull, TypeRegistry;
 
 /// A description of a SQL query as interpreted by this package.
 ///